requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
cffi==1.15.0
chardet==4.0.0
idna==3.3
multidict==6.0.2
py-cord==2.0.0b7
pyaml==21.10.1
pycparser==2.21
PyNaCl==1.5.0
PyYAML==6.0
typing_extensions==4.2.0
yarl==1.7.2
